backend_ops/local-api-obsidian_vault/integrations/tool-box/motia-docs-scraper/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Scraping Engines
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
playwright==1.40.0
scrapfly-sdk==1.3.0
selenium==4.15.2

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
yugabyte-psycopg2==2.9.9

# MCP Integration
mcp==1.0.0
context7-sdk==0.1.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
numpy==1.25.2
aiofiles==23.2.1
httpx==0.25.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Performance
aiohttp==3.9.1
asyncio-throttle==1.0.2
aioredis==2.0.1
uvloop==0.19.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0